Blocks implemented (not fully)
Now you can add blocks and mario will be able to walk on them. Only thing I would like to fix is that now, when he goes to a block that is higher than him, the game auto jumps him up to the platform. I would like to make it so that it doesn't do anything and just restricts movement so that they can't move.