Added Sounds
Added all sound effects that will be needed for this project. Note that these sounds will probably need to be trimmed as there is some time before the sound effect where nothing is playing and that needs to go in order for the animation and the sound effect to line up. So if anyone wants to trim those sounds please do so.