Beta Release
Fixed all major bugs. Now Mario is able to move underneath floating blocks, his jump is restricted by floating blocks, and his movements is restricted by blocks. One thing that needs to be done is the resizing of the images. All of Mario's images still need to be resized to be just a little bit smaller so that the grid layout works.