Fixed sound and finished first level
All sound clips have been trimmed now so that its just the sound that will play and no emtpy space. Also, the text reader has been implemented and now the first level has been rendered! With that accomplishment has also come a toooon of bugs. Now mario's movement gets kind of screwed up at times and occassionally he can go inside of blocks. It seems though that the only ones that he is going inside of are the stair blocks so I think it has to do with how many blocks are in a staircase and how they are all rendered.