Big Update
Fixed a bunch of bugs that had to do with movement, such as clipping through blocks and getting stuck in them as well. Next thing is pretty big. I have added the Prop interface which can be used for anything that doesnt have a rectangular collider and can also be used to draw things that aren't obstructive but just general background pieces.I have also added the images for a couple more things that can be added into the game such as the ending flagpole and the castle image at the end of a level as well. I have also added the pipes and their unique colliders into the game. At this point the first level's obstuctions are completely done.